+++ src/linad99/ivsort2.cpp
@@ -24,7 +24,7 @@ ivector sort(const ivector& v, int NSTACK)
    const int  FA=211;
    const int  FC=1663;
 
-  int n=v.size();
+  int n=static_cast<int>(v.size());
   ivector arr(v.indexmin(),v.indexmax());
   arr=v;
   arr.shift(1);
@@ -123,7 +123,7 @@ ivector sort(const ivector& _v, const ivector& _index, int NSTACK)
       << " in ivector sort(const ivector& v, const ivector& index)\n";
     ad_exit(1);
   }
-  int n=v.size();
+  int n=static_cast<int>(v.size());
   int One=1;
   int nne=v.indexmin();
   index.fill_seqadd(nne,One);
